+++ ChangeLog.md
@@ -2,7 +2,8 @@
 
 
 ## Unreleased changes
-* Added `InterpretersFor` as a shorthand for interpreters consuming multiple effects
+- Added `InterpretersFor` as a shorthand for interpreters consuming multiple effects
+- Added `runTSimple` and `bindTSimple`, which are simplified variants of `runT` and `bindT`
 
 ## 1.4.0.0 (2020-10-31)